% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{lunar-package}
\alias{lunar-package}
\title{The lunar Package}
\description{
Lunar Phase & Distance, Seasons and Other Environmental Factors.

Provides functions to calculate the phase of the moon,
  its distance from the earth, the season and possibly
  other environmental factors, based on date and location.
}
\details{
\tabular{ll}{
Package: \tab lunar\cr
Type: \tab Package\cr
Version: \tab 0.1-04\cr
Date: \tab 2014-09-04\cr
Depends: \tab R (>= 2.10.0)\cr
Encoding: \tab UTF-8\cr
License: \tab MIT\cr
LazyLoad: \tab no\cr
URL: \tab http://statistics.lazaridis.eu\cr
}

Provides functions to calculate lunar and other environmental covariates.

This package is used by the author to calculate covariates
  for studies of lunar effect on health and healthcare.

  References forthcoming.
}
\author{
Emmanuel Lazaridis
}
\keyword{package}

