% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lunar.package.R
\docType{data}
\name{lunar.8phases}
\alias{lunar.8phases}
\title{Lunar Phase Categories (8)}
\format{
An object of class \code{character} of length 8.
}
\usage{
lunar.8phases
}
\description{
Return 8 category labels for lunar phases.
}
\details{
These are category names corresponding to phases of
  the moon.
  Moon phase category names may be returned in
  the output of the
  \code{\link{lunar.phase}} function if its
  \code{name} option is set to \code{TRUE}.
}
\examples{
print(lunar.8phases)
}
\seealso{
\code{\link{lunar.phase}}
}
\keyword{lunar}
\keyword{moon}
\keyword{phase}
