% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lunar.package.R
\docType{data}
\name{lunar.distances}
\alias{lunar.distances}
\title{Lunar Distance Categories}
\format{
An object of class \code{character} of length 5.
}
\usage{
lunar.distances
}
\description{
Return category labels for lunar distances.
}
\details{
These are category names corresponding to distances
  from the center of the Earth to the center of its moon.
  Distance category names are used in the output of the
  \code{\link{lunar.distance}} function if its
  \code{name} option is set to \code{TRUE}.

  The perigee occurs at roughly \eqn{56} Earth radii, and
  the apogee at about \eqn{62.8} Earth radii.
  These categories are not determined according to any
  common standard.
  They may have different precise definitions
  for their bounds in different analyses.
}
\examples{
print(lunar.distances)
}
\seealso{
\code{\link{lunar.distance}}
}
\keyword{distance}
\keyword{lunar}
\keyword{moon}
