% Generated by roxygen2 (4.0.1): do not edit by hand
\name{terrestrial.season}
\alias{terrestrial.season}
\title{Terrestrial Season}
\usage{
terrestrial.season(x)
}
\arguments{
\item{x}{A vector of \code{\link[base]{Date}} values.}
}
\description{
Returns the season on specified dates.
}
\details{
The definitions for non-leap years are as follows
  (dates are inclusive):
  \describe{
    \item{Winter:}{21 December through 21 March}
    \item{Spring:}{22 March through 21 June}
    \item{Summer:}{22 June through 20 September}
    \item{Autumn:}{21 September through 20 December}
  }
  In leap years spring comes a day early!
}
\examples{
terrestrial.season(as.Date("2004-03-24"))
}
\seealso{
\code{\link{terrestrial.seasons}}
}
\keyword{earth}
\keyword{season}

