% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{terrestrial.seasons}
\alias{terrestrial.seasons}
\title{Terrestrial Season Categories}
\format{\preformatted{ chr [1:4] "Winter" "Spring" "Summer" "Autumn"
}}
\usage{
terrestrial.seasons
}
\description{
Return category labels for the seasons on Earth.
}
\details{
These are category names corresponding to the
  seasons of the planet Earth.
}
\examples{
print(terrestrial.seasons)
}
\keyword{earth}
\keyword{season}

