% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lunar-package.R, R/lunar.package.R
\docType{package}
\name{lunar-package}
\alias{lunar-package}
\title{The lunar Package}
\description{
Lunar Phase & Distance, Seasons and Other Environmental Factors.

Provides functions to calculate the phase of the moon,
  its distance from the earth, the season and possibly
  other environmental factors, based on date and location.
}
\details{
\tabular{ll}{
Package: \tab lunar\cr
Type: \tab Package\cr
Version: \tab 0.2-01\cr
Date: \tab 2022-06-13\cr
Depends: \tab R (>= 2.10.0)\cr
Encoding: \tab UTF-8\cr
License: \tab MIT\cr
LazyLoad: \tab no\cr
}

Provides functions to calculate lunar and other environmental covariates.

This package is used by the author to calculate covariates
  for studies of lunar effect on health and healthcare.
  It is based on an astronomical calculation that considers
  the moon's phase and its distance from the earth to calculate
  theoretical relative amount of moonlight on any given night.
  The package is not based on luminance data.
  It does not consider local cloud cover, for example, so it
  only provides maximum possible moonlight, not actual moonlight.

  References forthcoming.
}
\author{
Emmanuel Lazaridis
}
\keyword{package}
