% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{lunar.4phases}
\alias{lunar.4phases}
\title{Lunar Phase Categories (4)}
\format{\preformatted{ chr [1:4] "New" "Waxing" "Full" "Waning"
}}
\usage{
lunar.4phases
}
\description{
Return 4 category labels for lunar phases.
}
\details{
These are category names corresponding to phases of
  the moon.
  Moon phase category names may be returned in
  the output of the
  \code{\link{lunar.phase}} function if its
  \code{name} option is set to \code{TRUE}.
}
\examples{
print(lunar.4phases)
}
\seealso{
\code{\link{lunar.phase}}
}
\keyword{lunar}
\keyword{moon}
\keyword{phase}

