% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{lunar.8phases}
\alias{lunar.8phases}
\title{Lunar Phase Categories (8)}
\format{\preformatted{ chr [1:8] "New" "Waxing crescent" "First quarter" "Waxing gibbous" "Full" "Waning gibbous" ...
}}
\usage{
lunar.8phases
}
\description{
Return 8 category labels for lunar phases.
}
\details{
These are category names corresponding to phases of
  the moon.
  Moon phase category names may be returned in
  the output of the
  \code{\link{lunar.phase}} function if its
  \code{name} option is set to \code{TRUE}.
}
\examples{
print(lunar.8phases)
}
\seealso{
\code{\link{lunar.phase}}
}
\keyword{lunar}
\keyword{moon}
\keyword{phase}

