% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lunar.package.R
\docType{data}
\name{terrestrial.seasons}
\alias{terrestrial.seasons}
\title{Terrestrial Season Categories}
\format{
An object of class \code{character} of length 4.
}
\usage{
terrestrial.seasons
}
\description{
Return category labels for the seasons on Earth.
}
\details{
These are category names corresponding to the
  seasons of the planet Earth.
}
\examples{
print(terrestrial.seasons)
}
\keyword{earth}
\keyword{season}
